#95 #132 #133 refactored feature to the quick attachment drawer so that the camera direction icon flips when the camera direction
